Rename imaging Providers
